Apply suggestions from code review

+++ modules/nf-core/bismark/coverage2cytosine/main.nf
@@ -25,7 +25,9 @@ process BISMARK_COVERAGE2CYTOSINE {
     def args = task.ext.args ?: ''
     def prefix = task.ext.prefix ?: "${meta.id}"
     """
-    ln -s \$(readlink $fasta) $index/$fasta
+    if [ ! -f $index/$fasta ]; then
+        ln -s \$(readlink $fasta) $index/$fasta;
+    fi
 
     coverage2cytosine \\
         $coverage_file \\
@@ -33,8 +35,6 @@ process BISMARK_COVERAGE2CYTOSINE {
         --output ${prefix} \\
         --gzip \\
         $args
-
-    rm $index/$fasta
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
